agents/requirements.txt
anthropic>=0.86.0
web3>=7.0.0
python-dotenv>=1.0.0
